render table headers as buttons
