+++ vtv/generic/NameEdit.jsx
@@ -1,0 +1,87 @@
+import { useState, useRef, useEffect } from 'react'
+import setCaretAtEnd from '../util/setCaretAtEnd'
+import escapeHtml from '../util/escapeHtml'
+
+export default React.forwardRef(function NameEditInPlace(
+  { name, editingName, path, context: { onMessage, theme } },
+  ref
+) {
+  const editableRef = useRef(null)
+  const [newName, setNewName] = useState(editingName === 'new' ? '' : name)
+
+  useEffect(() => {
+    if (editableRef.current) {
+      editableRef.current.focus()
+      setCaretAtEnd(editableRef.current)
+    }
+  }, [editableRef])
+
+  const sendAction = (data = {}) => {
+    onMessage({
+      path,
+      action: 'rename',
+      editing: false,
+      ...data,
+    })
+  }
+
+  const save = (data = {}) => {
+    let value
+    const text = editableRef.current.innerText
+    setNewName(text)
+    try {
+      value = JSON.parse(text)
+    } catch (e) {
+      value = text
+    }
+    sendAction({
+      value,
+      ...data,
+    })
+  }
+
+  const cancel = () => {
+    sendAction()
+  }
+
+  const handleKeyPress = e => {
+    if ((e.key === 'Enter' && e.shiftKey === false) || e.key === 'Tab') {
+      e.preventDefault()
+      save({ tab: e.key === 'Tab', enter: e.key === 'Enter' })
+    } else if (e.key === 'Esc' || e.key === 'Escape') {
+      e.preventDefault()
+      cancel()
+    }
+    setNewName(editableRef.current.innerText)
+  }
+
+  return (
+    <div>
+      &#x200b;
+      <span
+        contentEditable
+        ref={editableRef}
+        onKeyDown={handleKeyPress}
+        onBlur={save}
+        dangerouslySetInnerHTML={{
+          __html: escapeHtml(editingName === 'new' ? '' : name),
+        }}
+      ></span>
+      <style jsx>{`
+        div {
+          background-color: ${theme.bubble1};
+          border-radius: 12px;
+          outline: none;
+          padding: 3px 7px;
+          border: 0;
+          outline: none;
+          margin: 0 0;
+        }
+        span {
+          outline: none;
+          margin: 0 0;
+        }
+      `}</style>
+    </div>
+  )
+})
