+++ vtv/generic/StringView.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 
 export default ({ value, maxLength }) => {
-  if (/\n|"|^$/.test(value)) {
+  if (/\n|\t|"|^$/.test(value)) {
     if (value.length > maxLength - 10) {
       return <em>{JSON.stringify(value.substr(0, maxLength - 10))}</em>
     } else {
