+++ vtv/table/ColumnHeaderView.jsx
@@ -1,0 +1,5 @@
+import NameButton from '../generic/NameButton'
+
+export default function ColumnHeaderView({ children, context }) {
+  return <NameButton context={context}>{children}</NameButton>
+}
