+++ vtv/table/RowHeaderView.jsx
@@ -1,0 +1,5 @@
+import NameButton from '../generic/NameButton'
+
+export default function RowHeaderView({ children, context }) {
+  return <NameButton context={context}>{children}</NameButton>
+}
