+++ vtv/tree/NodeNameView.jsx
@@ -1,122 +1,13 @@
-import React, { useState, useRef, useEffect } from 'react'
-import { Manager, Reference, Popper } from 'react-popper'
-import Textarea from '../generic/Textarea'
-import StringView from '../content/StringView'
+import { useState } from 'react'
+import { Manager, Reference } from 'react-popper'
 import NodeMenu from '../tree/NodeMenu'
-import setCaretAtEnd from '../util/setCaretAtEnd'
-import escapeHtml from '../util/escapeHtml'
+import NameButton from '../generic/NameButton'
+import NameEdit from '../generic/NameEdit'
+import StringView from '../generic/StringView'
 
 // Node Name Key
 
-const NameEditInPlace = React.forwardRef(
-  ({ name, editingName, path, context: { onMessage, theme } }, ref) => {
-    const editableRef = useRef(null)
-    const [newName, setNewName] = useState(editingName === 'new' ? '' : name)
-
-    useEffect(() => {
-      if (editableRef.current) {
-        editableRef.current.focus()
-        setCaretAtEnd(editableRef.current)
-      }
-    }, [editableRef])
-
-    const sendAction = (data = {}) => {
-      onMessage({
-        path,
-        action: 'rename',
-        editing: false,
-        ...data,
-      })
-    }
-
-    const save = (data = {}) => {
-      let value
-      const text = editableRef.current.innerText
-      setNewName(text)
-      try {
-        value = JSON.parse(text)
-      } catch (e) {
-        value = text
-      }
-      sendAction({
-        value,
-        ...data,
-      })
-    }
-
-    const cancel = () => {
-      sendAction()
-    }
-
-    const handleKeyPress = e => {
-      if ((e.key === 'Enter' && e.shiftKey === false) || e.key === 'Tab') {
-        e.preventDefault()
-        save({ tab: e.key === 'Tab', enter: e.key === 'Enter' })
-      } else if (e.key === 'Esc' || e.key === 'Escape') {
-        e.preventDefault()
-        cancel()
-      }
-      setNewName(editableRef.current.innerText)
-    }
-
-    return (
-      <div>
-        &#x200b;
-        <span
-          contentEditable
-          ref={editableRef}
-          onKeyDown={handleKeyPress}
-          onBlur={save}
-          dangerouslySetInnerHTML={{
-            __html: escapeHtml(editingName === 'new' ? '' : name),
-          }}
-        ></span>
-        <style jsx>{`
-          div {
-            background-color: ${theme.bubble1};
-            border-radius: 12px;
-            outline: none;
-            padding: 3px 7px;
-            border: 0;
-            outline: none;
-            margin: 0 0;
-          }
-          span {
-            outline: none;
-            margin: 0 0;
-          }
-        `}</style>
-      </div>
-    )
-  }
-)
-
-const NameButton = ({ displayName, name, onClick, context: { theme } }) => {
-  return (
-    <button className="id" onClick={onClick} tabIndex="-1">
-      {typeof displayName !== 'undefined' ? (
-        displayName
-      ) : (
-        <StringView value={name} maxLength={40} />
-      )}
-      <style jsx>{`
-        button {
-          cursor: pointer;
-          background-color: ${theme.bubble1};
-          color: ${theme.foreground};
-          font-family: ${theme.fontFamily};
-          border-radius: 9999px;
-          outline: none;
-          padding: 3px 7px;
-          font-size: inherit;
-          border: 0;
-        }
-      `}</style>
-    </button>
-  )
-}
-
-export default ({
+export default function NodeNameView({
   name,
   displayName,
   editingName,
@@ -124,7 +15,7 @@ export default ({
   path,
   onMessage,
   nodeMenuProps,
-}) => {
+}) {
   const [menuOpen, setMenuOpen] = useState(false)
 
   const rename = () => {
@@ -133,7 +24,7 @@ export default ({
 
   if (editingName) {
     return (
-      <NameEditInPlace
+      <NameEdit
         name={name}
         editingName={editingName}
         path={path}
@@ -146,13 +37,14 @@ export default ({
         <Reference>
           {({ ref }) => (
             <div ref={ref}>
-              <NameButton
-                displayName={displayName}
-                name={name}
-                onClick={() => setMenuOpen(true)}
-                onDoubleClick={rename}
-                context={context}
-              />
+              <NameButton onClick={() => setMenuOpen(true)} context={context}>
+                <StringView
+                  value={
+                    typeof displayName !== 'undefined' ? displayName : name
+                  }
+                  maxLength={60}
+                />
+              </NameButton>
             </div>
           )}
         </Reference>
